requirements.txt
google-generativeai
streamlit
requests